requirements.txt
asgiref==3.5.2
Django==4.1.3
django-admin-persian-fonts==0.2
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
persiantools==3.0.1
Pillow==9.3.0
PyJWT==2.6.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
